+++ frontend/src/services/graphql/Shared/getEnumByName.js
@@ -5,8 +5,8 @@ const getEnumByName = ({ enum_names }) => ({
     enum_names: enum_names
   },
   query: gql`
-    query queryEnum($enum_names: [String]) {
-      queryEnum(enums_names: $enum_names) {
+    query queryEnums($enum_names: [String]) {
+      queryEnums(enums_names: $enum_names) {
         name
         items {
           name
@@ -17,28 +17,15 @@ const getEnumByName = ({ enum_names }) => ({
   `
 });
 
-export const fetchOneEnum = async (client, enum_name) => {
-  const response = await client.query(
-    getEnumByName({ enum_names: [enum_name] })
-  );
-  if (!response.errors && response.data.queryEnum != null) {
-    return response.data.queryEnum[0].items;
-  } else {
-    return [];
-  }
-};
-
 export const fetchEnums = async (client, enum_names) => {
   const response = await client.query(
     getEnumByName({ enum_names: enum_names })
   );
-  let result = {};
-  if (!response.errors && response.data.queryEnum != null) {
-    for (let i = 0; i < response.data.queryEnum.length; i++) {
-      result[response.data.queryEnum[i].name] =
-        response.data.queryEnum[i].items;
-    }
-  } else {
-    return result;
+  if (!response.errors && response.data.queryEnums != null) {
+    return Object.assign(
+      {},
+      ...response.data.queryEnums.map((x) => ({ [x.name]: x.items }))
+    );
   }
+  return {};
 };
